Implement toString and iterator methods in line with API design
Change the functionality of toString and iterator methods
depending on the context in which a LODInstance was created.

If a LODInstance came from filtering on predicates of another
LODInstance, the iterator gives access to objects of the
matching statements.

If it wasn't created via filtering of another LODInstance, the
iterator returns the full statements.