Remove closing ?> PHP tag
If there is errant whitespace after this tag, it can
cause problems with adding additional output to buffers.

This can mess up apps which attempt to send a stream as
the response, as the length of the stream and the length
of the content sent in the response may differ.