Parallelise HTTP GETs
Provide a method to allow multiple URIs to be fetched in
parallel.

This is useful for when the graph needs to be populated quickly,
but we are going to inspect the RDF later, so we don't need
resources to be returned immediately.

Also extract the RDF loading code into a separate function to
make future testing easier.